Move build_domains_from_column_names to domain.py
